app/requirements.txt
celery==4.2.1
chardet==3.0.4
Django==2.1
django-cors-headers==2.2.0
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17
django-nine==0.1.13
django-rest-auth==0.9.3
djangorestframework==3.9.1
elasticsearch==6.2.0
elasticsearch-dsl==6.1.0
gunicorn==19.9
ipaddress==1.0.22
json-logging-py==0.2
mysqlclient==1.3.12
pdfminer.six==20170720
pycryptodome==3.6.1
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
urllib3==1.22
